third_party/WebKit/LayoutTests/editing/selection/select_all/select_all_textarea.html

<!doctype html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<textarea id="textarea"></textarea>
<script>
test(function() {
    var textarea = document.getElementById('textarea');
    textarea.focus();
    assert_false(document.queryCommandEnabled('SelectAll'));

    textarea.value = 'x';
    textarea.focus();
    assert_true(document.queryCommandEnabled('SelectAll'));
}, 'SelectAll is not enabled if textarea is empty');
</script>
